Added support for light control over Socket.IO
Utilizes the existing server for the web view, so the server has to be exposed to users. If one wanted to keep the web view closed, another approach would be needed.

Unfortunately, this implementation does not transfer binary data over Socket.IO. The reason is that the Socket.IO version (0.9.16) used in the project is too old, as binary support was introduced in Socket.IO 1.0 (https://socket.io/blog/introducing-socket-io-1-0/#Binary-support). Therefore, the version would have to be upgraded, which would likely require larger changes to the effect server.

A minor drawback is that I only used address instead of the full rinfo, since socket.io does not offer same fields. This could be unified more, though.